Move to QtPdf and add the Card class
Before the visualization used zoomable images. It was so slow, so now it is visualized with QtPdf, which is faster.
The Question class is now in a different file and the logic behind the cards changed. Now there is a vector of Cards. The new Card class is inherited by Question and PdfPage