Fix: Change card navigation
Before, the next flashcard button would not take into consideration the visualized card, but only the last flashcard visualized. In this way using the spinbox could have resorted to unexpected behaviors. For the flashcards, there was a similar situation.
Now it takes only the current visualized page into account.
There are two new variables in the Card class to memorize the previous PDF page and flashcard. In this way, the management of the pdf_visualization_model is easier