Feat: allow exports to Anki
It is possible to export the flashcards to a txt file that it is compatible with Anki. The txt file has to be imported manually in Anki and it is saved in data\Anki. The pdf pages related to a question are converted to jpg and saved in the media directory of Anki. The same relative path of the pdf is copied in the Anki media directory and in data\Anki