Fix: update pdf with less page
A new version of a pdf could have fewer pages, but there could be flashcards that are pointing to a page number that does not exist anymore. Now the visualization and the export manage also this condition. The new page number will be the last page, but the old reference page on file won't be changed until a new update/addition/deletion of a flashcard